code organize & refactors
