+++ compose_cfg/dev_humble_nvidia.yaml
@@ -1,28 +1,28 @@
-version: '3.8'
+version: "3.8"
 
 services:
-  robotics-academy:
-    image: jderobot/robotics-backend:latest
-    container_name: developer-container
-    deploy:
-      resources:
-        reservations:
-          devices:
-            - driver: nvidia
-              count: all
-              capabilities: [gpu]
-    runtime: nvidia
-    environment:
-      - NVIDIA_VISIBLE_DEVICES=all
-      - NVIDIA_DRIVER_CAPABILITIES=all
-    ports:
-      - "7164:7164"
-      - "7163:7163"
-      - "6080:6080"
-      - "1108:1108"
-    volumes:
-      - type: bind
-        source: ./
-        target: /RoboticsAcademy
-    tty: true
-    stdin_open: true
+    robotics-academy:
+        image: jderobot/robotics-backend:latest
+        container_name: developer-container
+        deploy:
+            resources:
+                reservations:
+                    devices:
+                        - driver: nvidia
+                          count: all
+                          capabilities: [gpu]
+        runtime: nvidia
+        environment:
+            - NVIDIA_VISIBLE_DEVICES=all
+            - NVIDIA_DRIVER_CAPABILITIES=all
+        ports:
+            - "7164:7164"
+            - "7163:7163"
+            - "6080:6080"
+            - "1108:1108"
+        volumes:
+            - type: bind
+              source: ./
+              target: /RoboticsAcademy
+        tty: true
+        stdin_open: true
