+++ react_frontend/src/components/editors/monaco-editor/autocomplete-snippets/import_snippets.js
@@ -1,52 +1,3 @@
-export const snippetsBuilder = ({ monaco, range, importName }) => {
-  const snippets = [];
-
-  const importSnippets = importSnippetsObj[`_${importName}`];
-
-  if (!importSnippets) return [];
-
-  importSnippets.forEach((snippet) => {
-    if (snippet.label && snippet.code) {
-      snippets.push({
-        label: snippet.label,
-        kind:
-          snippet.type === "method"
-            ? monaco.languages.CompletionItemKind.Method
-            : monaco.languages.CompletionItemKind.Variable,
-        insertText: snippet.code,
-        range: range,
-      });
-    }
-  });
-
-  return snippets;
-};
-
-export const extractPythonImports = (code) => {
-  const importRegex =
-    /\bimport\s+([a-zA-Z_][\w]*)(?:\s+as\s+([a-zA-Z_][\w]*))?/g;
-  // const fromImportRegex =
-  //     /\bfrom\s+([a-zA-Z_][\w]*)\s+import\s+([\w,*{}]+)(?:\s+as\s+([a-zA-Z_][\w]*))?/g;
-
-  let imports = [];
-  let match;
-
-  while ((match = importRegex.exec(code)) !== null) {
-    const importName = match[1];
-    const alias = match[2] ? match[2] : importName;
-    imports.push({ importName, alias });
-  }
-
-  // while ((match = fromImportRegex.exec(code)) !== null) {
-  //     const importFrom = match[1];
-  //     const importWhat = match[2];
-  //     const alias = match[3] ? match[3] : null;
-  //     imports.push({ importFrom, importWhat, alias });
-  // }
-
-  return imports;
-};
-
 export const importSnippetsObj = {
   _numpy: [
     {
